Move http client to library
HttpClient defined for testing purposes was moved to lib itself and is
available thgough Application.api_client. This is to make it available
generally.

In addiion ot that couple modification was made. The client has
pluggable authentication for example. It also keeps retry count
implemented as it was. To modify the behavior it one should pass
appropriate requests.Session.